+++ projetFinalDeMonCote/src/main/java/co/simplon/dao/ConclusionDAO.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import co.simplon.domain.Conclusion;
 
 @Repository
-public interface ConclusionDAO extends CrudRepository<Conclusion, Long>{
+public interface ConclusionDAO extends CrudRepository<Conclusion, Integer>{
 
 }
