+++ projetFinalDeMonCote/src/main/java/co/simplon/dao/QuestionDAO.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import co.simplon.domain.Question;
 
 @Repository
-public interface QuestionDAO extends CrudRepository<Question, Long>{
+public interface QuestionDAO extends CrudRepository<Question, Integer>{
 
 }
